Refactor the login context
